修复：通过账户用户统计的类型检查
将 API Key 计数转换为用户列表响应约定的 int64，并同步远端 gofmt 调整的导入顺序。

已在远端通过 go test ./...、Vue 类型检查、Vitest 与生产构建。

+++ internal/service/user/service.go
@@ -104,7 +104,7 @@ func (s *Service) List(ctx context.Context) ([]ManagedUser, error) {
 		if err != nil {
 			return nil, gerror.Wrap(err, "count user API keys")
 		}
-		result = append(result, ManagedUser{Profile: profileFromEntity(row), APIKeyCount: keyCount, Usage: summary})
+		result = append(result, ManagedUser{Profile: profileFromEntity(row), APIKeyCount: int64(keyCount), Usage: summary})
 	}
 	return result, nil
 }
